// src/app/cores/models/category.model.ts
export interface CategoryModel {
    categoryId: number;
    categoryName: string | null;
    isActive: boolean | null;
    categoryImage: string | null;
    //createdAt: string | null;
    //updatedAt: string | null;
    //deletedAt: string | null;
   // menus: Menu[];
}//categoryId,categoryName,isActive,categoryImage,createdAt,updatedAt,deletedAt